+++ lending/src/codat_lending/models/accounting_bill_payment_all_of_payment_method_ref.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from codat_lending.models.shared.paymentmethodref import PaymentMethodRef as AccountingBillPaymentAllOfPaymentMethodRef, PaymentMethodRefTypedDict as AccountingBillPaymentAllOfPaymentMethodRefTypedDict
+
+__all__ = ["AccountingBillPaymentAllOfPaymentMethodRef", "AccountingBillPaymentAllOfPaymentMethodRefTypedDict"]